+++ src/pages/admin/categories/edit/[id].astro
@@ -2,7 +2,7 @@
 export const prerender = false;
 
 import Layout from '../../../../layouts/Layout.astro';
-import CategoryEditForm from '../../../../components/admin/categories/CategoryEditForm';
+import CategoryEditForm from '../../../../features/admin/categories/components/CategoryEditForm';
 
 const { id } = Astro.params;
