+++ src/pages/admin/categories/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../../layouts/Layout.astro';
-import CategoryList from '../../../components/admin/categories/CategoryList';
+import CategoryList from '../../../features/admin/categories/components/CategoryList';
 ---
 
 <Layout title="Categorías · SansiStore Admin">
