+++ src/pages/admin/categories/new.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../../layouts/Layout.astro';
-import CategoryForm from '../../../components/admin/categories/CategoryForm';
+import CategoryForm from '../../../features/admin/categories/components/CategoryForm';
 ---
 
 <Layout title="Nueva categoría · SansiStore Admin">
